+++ metrictank/src/main/java/com/expedia/metrics/metrictank/MetricTankIdFactory.java
@@ -23,12 +23,8 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-import static com.expedia.metrics.metrictank.MessagePackSerializer.INTERVAL;
-import static com.expedia.metrics.metrictank.MessagePackSerializer.ORG_ID;
 
 public class MetricTankIdFactory implements IdFactory {
     
@@ -38,32 +34,15 @@ public class MetricTankIdFactory implements IdFactory {
     }
     
     public MetricKey getKey(MetricDefinition metric) {
-        Map<String, String> tags = new HashMap<>(metric.getTags().getKv());
-        final int orgId;
-        try {
-            orgId = Integer.parseInt(tags.remove(ORG_ID));
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("Tag 'org_id' must be an int", e);
-        }
+        final int orgId = MessagePackSerializer.getOrgId(metric);
         final String name = metric.getKey();
         if (name == null) {
             throw new IllegalArgumentException("Property 'key' is required by metrictank");
         }
-        final int interval;
-        try {
-            interval = Integer.parseInt(tags.remove(INTERVAL));
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("Tag 'interval' must be an int", e);
-        }
-        final String unit = tags.remove(MetricDefinition.UNIT);
-        if (unit == null) {
-            throw new IllegalArgumentException("Tag 'unit' is required by metrictank");
-        }
-        final String mtype = tags.remove(MetricDefinition.MTYPE);
-        if (mtype == null) {
-            throw new IllegalArgumentException("Tag 'mtype' is required by metrictank");
-        }
-        List<String> formattedTags = formatTags(tags);
+        final int interval = MessagePackSerializer.getInterval(metric);
+        final String unit = MessagePackSerializer.getUnit(metric);
+        final String mtype = MessagePackSerializer.getMtype(metric);
+        List<String> formattedTags = formatTags(metric.getTags().getKv());
         return getKey(orgId, name, unit, mtype, interval, formattedTags);
     }